requirements.txt
# OpenElara Backend Dependencies
# Python 3.8+ required

# RAG & Vector Database
chromadb>=0.4.0
langchain>=0.1.0
sentence-transformers>=2.2.0

# Document Processing
python-docx>=1.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
openpyxl>=3.1.0
pandas>=2.0.0

# Image Processing
Pillow>=10.0.0
pytesseract>=0.3.10

# Text Processing
markdownify>=0.11.0
chardet>=5.0.0
html2text>=2020.1.16

# Web Scraping
scrapy>=2.11.0

# Local LLM Integration
ollama>=0.1.0

# Optional: For better performance
# numpy>=1.24.0
# scipy>=1.10.0